restructured folders for test finalisation, common test finalisation components now stored in common components
