+++ src/components/test-finalisation/debrief-witnessed/debrief-witnessed.ts
@@ -3,7 +3,7 @@ import { FormGroup, FormControl, Validators } from '@angular/forms';
 import {
   OutcomeBehaviourMapProvider,
   VisibilityType,
-} from '../../../../providers/outcome-behaviour-map/outcome-behaviour-map';
+} from '../../../providers/outcome-behaviour-map/outcome-behaviour-map';
 
 enum ValidWitnessedValues {
   YES = 'Yes',
