+++ src/components/test-finalisation/test-finalisation-component.module.ts
@@ -1,11 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from 'ionic-angular';
-import { D255Component } from '../components/d255/d255';
-import { DebriefWitnessedComponent } from '../components/debrief-witnessed/debrief-witnessed';
-import { LanguagePreferencesComponent } from '../components/language-preference/language-preferences';
-import { FinalisationHeaderComponent } from '../components/finalisation-header/finalisation-header';
-import { ComponentsModule } from '../../../components/common/common-components.module';
+import { D255Component } from './d255/d255';
+import { DebriefWitnessedComponent } from './debrief-witnessed/debrief-witnessed';
+import { LanguagePreferencesComponent } from './language-preference/language-preferences';
+import { FinalisationHeaderComponent } from './finalisation-header/finalisation-header';
 
 @NgModule({
   declarations: [
@@ -16,7 +15,6 @@ import { ComponentsModule } from '../../../components/common/common-components.m
   ],
   imports: [
     CommonModule,
-    ComponentsModule,
     IonicModule,
   ],
   exports: [
