+++ src/pages/pass-finalisation/cat-b/_tests_/pass-finalisation.cat-b.page.spec.ts
@@ -1,18 +1,18 @@
 import { ComponentFixture, async, TestBed } from '@angular/core/testing';
 import { IonicModule, NavController, NavParams, Config, Platform } from 'ionic-angular';
 import { NavControllerMock, NavParamsMock, ConfigMock, PlatformMock } from 'ionic-mocks';
-import { AppModule } from '../../../../../app/app.module';
+import { AppModule } from '../../../../app/app.module';
 import { PassFinalisationCatBPage } from '../pass-finalisation.cat-b.page';
-import { AuthenticationProvider } from '../../../../../providers/authentication/authentication';
-import { AuthenticationProviderMock } from '../../../../../providers/authentication/__mocks__/authentication.mock';
-import { DateTimeProvider } from '../../../../../providers/date-time/date-time';
-import { DateTimeProviderMock } from '../../../../../providers/date-time/__mocks__/date-time.mock';
+import { AuthenticationProvider } from '../../../../providers/authentication/authentication';
+import { AuthenticationProviderMock } from '../../../../providers/authentication/__mocks__/authentication.mock';
+import { DateTimeProvider } from '../../../../providers/date-time/date-time';
+import { DateTimeProviderMock } from '../../../../providers/date-time/__mocks__/date-time.mock';
 import { Store } from '@ngrx/store';
-import { StoreModel } from '../../../../../shared/models/store.model';
-import { PersistTests } from '../../../../../modules/tests/tests.actions';
+import { StoreModel } from '../../../../shared/models/store.model';
+import { PersistTests } from '../../../../modules/tests/tests.actions';
 import { MockComponent } from 'ng-mocks';
-import { PracticeModeBanner } from '../../../../../components/common/practice-mode-banner/practice-mode-banner';
-import { TestFinalisationComponentsModule } from '../../../components/test-finalisation.module';
+import { PracticeModeBanner } from '../../../../components/common/practice-mode-banner/practice-mode-banner';
+import { TestFinalisationComponentsModule } from '../../../../components/test-finalisation/test-finalisation-component.module';
 
 describe('PassFinalisationPage', () => {
   let fixture: ComponentFixture<PassFinalisationCatBPage>;
