+++ src/pages/pass-finalisation/cat-b/pass-finalisation.module.cat-b.page.ts
@@ -3,8 +3,8 @@ import { IonicPageModule } from 'ionic-angular';
 import { PassFinalisationCatBPage } from './pass-finalisation.cat-b.page';
 import { EffectsModule } from '@ngrx/effects';
 import { PassFinalisationAnalyticsEffects } from '../pass-finalisation.analytics.effects';
-import { ComponentsModule } from '../../../../components/common/common-components.module';
-import { TestFinalisationComponentsModule } from '../../components/test-finalisation.module';
+import { ComponentsModule } from '../../../components/common/common-components.module';
+import { TestFinalisationComponentsModule } from '../../../components/test-finalisation/test-finalisation-component.module';
 
 @NgModule({
   declarations: [
